test_mount_stress: include single stress test
With this change we are including one stress test with near maximum values
i.e num_clients=20, num_operations=40, file_size=2**25

Other existing tests with lower values are eliminated as the goal
here should be to test with maximum stress. Tests with lower values are
needless when the test with maximum stress passes.

Fixes: #54
Signed-off-by: Shwetha K Acharya <Shwetha.K.Acharya@ibm.com>