pyproject.toml: register pytest mark privileged
f53d5d3 testcases: mark certain tests as privileged

Introduced the custom mark privileged to mark out tests which need to be
run as privileged processes. However this leads to warning messages
seen in the warnings summary for the tests.

We register these custom marks to avoid these warning messages.

Signed-off-by: Sachin Prabhu <sp@spui.uk>